Refactor scheduler
