+++ lib/application.js
@@ -9,7 +9,6 @@ const { MessageChannel, parentPort, threadId } = wt;
 const workerData = wt.workerData || { path: process.cwd() };
 const metavm = require('metavm');
 const metawatch = require('metawatch');
-const metautil = require('metautil');
 const { Interfaces } = require('./interfaces.js');
 const { Modules } = require('./modules.js');
 const { Resources } = require('./resources.js');
@@ -84,9 +83,9 @@ class Application extends events.EventEmitter {
     ]);
     this.sandbox.application.emit('loaded');
     await Promise.allSettled(this.starts.map((fn) => this.execute(fn)));
+    this.starts = [];
     this.sandbox.application.emit('started');
     await this.api.load();
-    if (this.kind === 'scheduler') await this.scheduler.load();
     const { api } = this.sandbox;
     if (api.auth) {
       const { provider } = api.auth;
@@ -98,7 +97,6 @@ class Application extends events.EventEmitter {
         api.auth.provider = provider;
       }
     }
-    this.starts = [];
     this.initialization = false;
     this.sandbox.application.emit('initialized');
   }
@@ -226,26 +224,4 @@ class Application extends events.EventEmitter {
   }
 }
 
-const application = new Application();
-
-if (parentPort) {
-  parentPort.on('message', async ({ name, exclusive, data, port }) => {
-    if (name !== 'invoke') return;
-    const { method, args } = data;
-    const handler = metautil.namespaceByPath(application.sandbox, method);
-    if (!handler) {
-      const error = new Error('Handler not found');
-      port.postMessage({ name: 'error', error });
-      return;
-    }
-    try {
-      const result = await handler(args);
-      port.postMessage({ name: 'invoke/done', data: result, exclusive });
-    } catch (error) {
-      port.postMessage({ name: 'error', error });
-      application.console.error(error.stack);
-    }
-  });
-}
-
-module.exports = application;
+module.exports = new Application();
