+++ lib/worker.js
@@ -7,6 +7,7 @@ const { Config } = require('metaconfiguration');
 const { Logger } = require('metalog');
 const { Server } = require('metacom');
 const metavm = require('metavm');
+const metautil = require('metautil');
 const { notLoaded } = require('./dependencies.js');
 const application = require('./application.js');
 
@@ -65,15 +66,38 @@ process.on('unhandledRejection', logError('unhandledRejection'));
     application.server = new Server(options, application);
   }
 
-  console.info(`Application started in worker ${threadId}`);
-  parentPort.postMessage({ name: 'started' });
+  const stop = async () => {
+    if (application.finalization) return;
+    console.info(`Graceful shutdown in worker ${threadId}`);
+    await application.shutdown();
+    process.exit(0);
+  };
 
-  parentPort.on('message', async ({ name }) => {
-    if (name === 'stop') {
-      if (application.finalization) return;
-      console.info(`Graceful shutdown in worker ${threadId}`);
-      await application.shutdown();
-      process.exit(0);
+  const invoke = async ({ exclusive, data, port }) => {
+    const { method, args } = data;
+    const handler = metautil.namespaceByPath(application.sandbox, method);
+    if (!handler) {
+      const error = new Error('Handler not found');
+      port.postMessage({ name: 'error', error });
+      return;
     }
+    try {
+      const result = await handler(args);
+      const msg = { name: 'invoke', status: 'done', data: result, exclusive };
+      port.postMessage(msg);
+    } catch (error) {
+      port.postMessage({ name: 'error', error });
+      application.console.error(error.stack);
+    }
+  };
+
+  const handlers = { stop, invoke };
+
+  parentPort.on('message', async (msg) => {
+    const handler = handlers[msg.name];
+    if (handler) handler(msg);
   });
+
+  console.info(`Application started in worker ${threadId}`);
+  parentPort.postMessage({ name: 'started', kind });
 })().catch(logError(`Can not start Application in worker ${threadId}`));
